added new bundle to main page
