+++ assets/js/bundles/actions.js
@@ -1,0 +1,1 @@
+require('../getEtlLastUpdate');
