+++ server/wqp/portal_ui_blueprint/templates/index.html
@@ -882,5 +882,6 @@
 {% endblock %}
 
 {% block body_page_script %}
-  <script type="text/javascript" src="{{ 'scripts/portal.js' | asset_url }}"></script>
+    <script type="text/javascript" src="{{ 'scripts/actions.js' | asset_url }}"></script>
+    <script type="text/javascript" src="{{ 'scripts/portal.js' | asset_url }}"></script>
 {% endblock %}
